fix: Better detection and handling of unhealthy mcp connections
This change moves most of the connection health logic out of backend_http_burp.go and into the mcp package.
When created a ping every 4 second starts, if it detects an unhealthy state it will mark the connection as unhealthy so that the next request will establish a new one.
Additionally retry logic is improved, and no longer done on context timeouts / deadline (respecting client configured timeouts.